pyproject.toml
[project]
name = "myproject"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
    "django (>=5.1,<6.0)",
    "psycopg[binary] (>=3.2,<4.0)",
    "uvicorn[standard] (>=0.32,<0.33)",
    "gunicorn (>=23.0,<24.0)",
    "celery (>=5.4,<6.0)",
    "redis (>=5.0,<6.0)",
    "django-environ (>=0.11,<0.12)",
    "django-htmx (>=1.27.0,<2.0.0)",
    "whitenoise (>=6.11.0,<7.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-django = "^4.9"
pytest-asyncio = "^0.24"
ruff = "^0.8"
pytest-playwright = "^0.7.2"
pytest-xdist = "^3.8.0"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.local"
python_files = ["test_*.py", "*_test.py"]
asyncio_mode = "auto"
addopts = "-n 8 --dist loadfile"

[tool.ruff]
line-length = 100
target-version = "py312"
exclude = ["**/migrations/*.py"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
